requirements.txt
asgiref==3.5.2
dj-database-url==0.5.0
dj-static==0.0.6
Django==4.0.4
django-ckeditor==6.4.2
django-crispy-forms==1.14.0
django-js-asset==2.0.0
gunicorn==20.1.0
requests==2.27.1
sqlparse==0.4.2
tzdata==2022.1
whitenoise==6.2.0
